Extract abstract class
