+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/DefaultRequestExceptionHandler.java
@@ -17,43 +17,14 @@
 package com.palantir.atlasdb.keyvalue.cassandra;
 
 import java.net.InetSocketAddress;
-import java.net.SocketException;
-import java.net.SocketTimeoutException;
-import java.time.Duration;
-import java.util.NoSuchElementException;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.function.Supplier;
 
-import org.apache.cassandra.thrift.InvalidRequestException;
-import org.apache.cassandra.thrift.TimedOutException;
-import org.apache.cassandra.thrift.UnavailableException;
-import org.apache.thrift.transport.TTransportException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.slf4j.helpers.MessageFormatter;
-
-import com.google.common.annotations.VisibleForTesting;
-import com.palantir.atlasdb.keyvalue.api.InsufficientConsistencyException;
-import com.palantir.atlasdb.keyvalue.cassandra.CassandraClientFactory.ClientCreationFailedException;
-import com.palantir.logsafe.SafeArg;
-import com.palantir.logsafe.UnsafeArg;
-
-class DefaultRequestExceptionHandler {
-    private static final Logger log = LoggerFactory.getLogger(CassandraClientPool.class);
-
-    private static final Duration backoffDuration = Duration.ofSeconds(1);
-
-    private final Supplier<Integer> maxTriesSameHost;
-    private final Supplier<Integer> maxTriesTotal;
-    private final Blacklist blacklist;
-
+class DefaultRequestExceptionHandler extends AbstractRequestExceptionHandler {
     DefaultRequestExceptionHandler(
             Supplier<Integer> maxTriesSameHost,
             Supplier<Integer> maxTriesTotal,
             Blacklist blacklist) {
-        this.maxTriesSameHost = maxTriesSameHost;
-        this.maxTriesTotal = maxTriesTotal;
-        this.blacklist = blacklist;
+        super(maxTriesSameHost, maxTriesTotal, blacklist);
     }
 
     @SuppressWarnings("unchecked")
@@ -81,140 +52,4 @@ class DefaultRequestExceptionHandler {
         handleBackoff(req, hostTried, ex);
         handleRetryOnDifferentHosts(req, hostTried, ex);
     }
-
-    @SuppressWarnings("unchecked")
-    private <K extends Exception> void logAndThrowException(int numberOfAttempts, Exception ex) throws K {
-        if (ex instanceof TTransportException
-                && ex.getCause() != null
-                && (ex.getCause().getClass() == SocketException.class)) {
-            log.error(CONNECTION_FAILURE_MSG, numberOfAttempts, ex);
-            String errorMsg =
-                    MessageFormatter.format(CONNECTION_FAILURE_MSG, numberOfAttempts).getMessage();
-            throw (K) new TTransportException(((TTransportException) ex).getType(), errorMsg, ex);
-        } else {
-            log.error("Tried to connect to cassandra {} times.",
-                    SafeArg.of("numTries", numberOfAttempts), ex);
-            throw (K) ex;
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private <K extends Exception> void logNumberOfAttempts(Exception ex,
-            int numberOfAttempts) throws K {
-        // Only log the actual exception the first time
-        if (numberOfAttempts > 1) {
-            log.info("Error occurred talking to cassandra. Attempt {} of {}. Exception message was: {} : {}",
-                    SafeArg.of("numTries", numberOfAttempts),
-                    SafeArg.of("maxTotalTries", maxTriesTotal),
-                    SafeArg.of("exceptionClass", ex.getClass().getTypeName()),
-                    UnsafeArg.of("exceptionMessage", ex.getMessage()));
-        } else {
-            log.info("Error occurred talking to cassandra. Attempt {} of {}.",
-                    SafeArg.of("numTries", numberOfAttempts),
-                    SafeArg.of("maxTotalTries", maxTriesTotal),
-                    ex);
-        }
-    }
-
-    private <K extends Exception> void handleBackoff(RetryableCassandraRequest<?, K> req,
-            InetSocketAddress hostTried,
-            Exception ex) {
-        if (!shouldBackoff(ex)) {
-            return;
-        }
-
-        int numberOfAttempts = req.getNumberOfAttempts();
-        long backoffPeriod = backoffDuration.toMillis();
-        // And value between -500 and +500ms to backoff to better spread load on failover
-        long sleepDuration =
-                numberOfAttempts * backoffPeriod + (ThreadLocalRandom.current().nextInt(1000) - 500);
-        log.info("Retrying a query, {}, with backoff of {}ms, intended for host {}.",
-                UnsafeArg.of("queryString", req.getFunction().toString()),
-                SafeArg.of("sleepDuration", sleepDuration),
-                SafeArg.of("hostName", CassandraLogHelper.host(hostTried)));
-
-        try {
-            Thread.sleep(sleepDuration);
-        } catch (InterruptedException i) {
-            throw new RuntimeException(i);
-        }
-    }
-
-    private <K extends Exception> void handleRetryOnDifferentHosts(RetryableCassandraRequest<?, K> req,
-            InetSocketAddress hostTried, Exception ex) {
-        if (shouldRetryOnDifferentHost(ex, req.getNumberOfAttempts())) {
-            log.info("Retrying with on a different host a query intended for host {}.",
-                    SafeArg.of("hostName", CassandraLogHelper.host(hostTried)));
-            req.giveUpOnPreferredHost();
-        }
-    }
-
-    // Determine the behavior we want from each type of exception.
-
-    @VisibleForTesting
-    boolean isRetryable(Exception ex) {
-        return isConnectionException(ex)
-                || isTransientException(ex)
-                || isIndicativeOfCassandraLoad(ex)
-                || isFastFailoverException(ex);
-    }
-
-    @VisibleForTesting
-    boolean shouldBlacklist(Exception ex, int numberOfAttempts) {
-        return isConnectionException(ex) && numberOfAttempts >= maxTriesSameHost.get();
-    }
-
-    @VisibleForTesting
-    boolean shouldBackoff(Exception ex) {
-        return isConnectionException(ex) || isIndicativeOfCassandraLoad(ex);
-    }
-
-    @VisibleForTesting
-    boolean shouldRetryOnDifferentHost(Exception ex, int numberOfAttempts) {
-        return isFastFailoverException(ex)
-                || (numberOfAttempts >= maxTriesSameHost.get()
-                && (isConnectionException(ex) || isIndicativeOfCassandraLoad(ex)));
-    }
-
-    // Group exceptions by type.
-
-    static boolean isConnectionException(Throwable ex) {
-        return ex != null
-                // tcp socket timeout, possibly indicating network flake, long GC, or restarting server.
-                && (ex instanceof SocketTimeoutException
-                || ex instanceof ClientCreationFailedException
-                || isConnectionException(ex.getCause()));
-    }
-
-    private static boolean isTransientException(Throwable ex) {
-        return ex != null
-                // There's a problem with the connection to Cassandra.
-                && (ex instanceof TTransportException
-                // Cassandra timeout. Maybe took too long to CAS, or Cassandra is under load.
-                || ex instanceof TimedOutException
-                // Not enough Cassandra nodes are up.
-                || ex instanceof InsufficientConsistencyException
-                || isTransientException(ex.getCause()));
-    }
-
-    private boolean isIndicativeOfCassandraLoad(Throwable ex) {
-        return ex != null
-                // pool for this node is fully in use
-                && (ex instanceof NoSuchElementException
-                // remote cassandra node couldn't talk to enough other remote cassandra nodes to answer
-                || ex instanceof UnavailableException
-                || isIndicativeOfCassandraLoad(ex.getCause()));
-    }
-
-    private boolean isFastFailoverException(Throwable ex) {
-        return ex != null
-                // underlying cassandra table does not exist. The table might exist on other cassandra nodes.
-                && (ex instanceof InvalidRequestException
-                || isFastFailoverException(ex.getCause()));
-    }
-
-    private static final String CONNECTION_FAILURE_MSG = "Tried to connect to cassandra {} times."
-            + " Error writing to Cassandra socket."
-            + " Likely cause: Exceeded maximum thrift frame size;"
-            + " unlikely cause: network issues.";
 }
